Add files via upload
The information is presented in raw in CVS format, the Ministry of Health of Mexico since April 14, 2020 published the cases associated with (SARS-CoV-2) COVID-19 as open data. The data processing corresponds to the records on the epidemic (SARS-CoV-2) COVID-19 at 1  June 2020. The treatment of the information is carried out through the application software for data mining Orange version 3.25.1, in which the algorithm for the analysis of information is filtered to present the current scenario in Mexico of the SARS-CoV-2 (COVID 19).
Government of Mexico. Health Secretary. Databases Covid-19 México 
https://datos.gob.mx/busca/dataset/informacion-referente-a-casos-covid-19-en-mexico/resource/e8c7079c-dc2a-4b6e-8035-08042ed37165